requirements.txt
pyTelegramBotAPI==4.14.0
openai>=1.55.0
python-dotenv==1.0.0
opencv-python>=4.8
numpy>=2.2
Pillow>=10.0